% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_add_frame.R
\name{add_frame}
\alias{add_frame}
\title{Add frame to ggplot2}
\usage{
add_frame(color = "black", linewidth = 1)
}
\arguments{
\item{color}{color to use for borders}

\item{linewidth}{widht of borders}
}
\value{
ggplot2 element/object
}
\description{
adds border to all sides of the plot
}
