% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_add_scalebar.R
\name{add_scalebar}
\alias{add_scalebar}
\title{Add scalebar}
\usage{
add_scalebar(
  xpos = 200,
  size = 2000,
  ypos = 200,
  height = 50,
  unit = "µm",
  color = "white",
  fontsize = 3.88
)
}
\arguments{
\item{xpos}{x-position to start the scalebar}

\item{size}{size (width) of scalebar}

\item{ypos}{y-position to start the scalebar}

\item{height}{height of scalebar}

\item{unit}{character string for the unit to add to the scalebar}

\item{color}{color of scalebar}

\item{fontsize}{fontsize of text used for scalebar}
}
\value{
ggplot2 annotation
}
\description{
Adds scalebar to ggplot2 object
}
