% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coloc_visium.R
\name{build_adjacency_table}
\alias{build_adjacency_table}
\title{Build adjacency table}
\usage{
build_adjacency_table(spot_pos_type)
}
\arguments{
\item{spot_pos_type}{dataframe with spot positions + info}
}
\value{
adjacency matrix
}
\description{
Build adjacency table
}
\references{
Adapted from Greenwald 2024 (https://github.com/tiroshlab/Spatial_Glioma)
}
