% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_build_plot.R
\name{build_clustree_plot}
\alias{build_clustree_plot}
\title{Build clustree plot}
\usage{
build_clustree_plot(plist, legend)
}
\arguments{
\item{plist}{names list of clustree plots, names are the genes}

\item{legend}{common legend to add, containing the cluster_size, sample count and In-propotion legend}
}
\value{
cowplot grid with plots
}
\description{
Combine clustrees with overlay of expression of multiple genes that share a common legend.
}
