% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_spatial.R
\name{compute_neighbor_metrics}
\alias{compute_neighbor_metrics}
\title{Compute neighbor metrics}
\usage{
compute_neighbor_metrics(
  approach,
  dbscan_res,
  group_varname,
  df_with_labels,
  n_cores = 1,
  k_neighbors = NA
)
}
\arguments{
\item{approach}{'knn' or 'radius'}

\item{group_varname}{variable containing your labels for the cell_ids, e.g. cell type}

\item{df_with_labels}{dataframe with 'cell_ids' and your 'group_varname'}

\item{n_cores}{number of cores to use}

\item{k_neighbors}{number of neighbors when approach = 'knn'}

\item{cell_id}{cell_id for which you want to assess its neighbors}
}
\value{
return dataframe with the columns: labels_of_neighbors, cell_id, n_total, and n_cells
}
\description{
Determine the number of neighbors per cell type for a cell of interest
}
