% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_compute_total_facetgrid_dimensions.R
\name{compute_nrows_in_facetgrid}
\alias{compute_nrows_in_facetgrid}
\title{Determine no. rows in facet grid}
\usage{
compute_nrows_in_facetgrid(n_plots, n_plots_per_row = 5)
}
\arguments{
\item{n_plots}{number of total facets (plot)}

\item{n_plots_per_row}{number of plots per row, i.e. number of columns}
}
\value{
numerical value
}
\description{
Determine no. rows in facet grid
}
