% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_spatial.R
\name{compute_perm_pval}
\alias{compute_perm_pval}
\title{Compute p-val permutation testing}
\usage{
compute_perm_pval(celltype_pair_oi, shuffled_df, obs_df, metric)
}
\arguments{
\item{celltype_pair_oi}{string with cell type pair of interest existing in shuffled_df and obs_df}

\item{shuffled_df}{dataframe with 'source_target' and 'metric' for all iterations (permutations)}

\item{obs_df}{dataframe with 'source_target' and 'metric' (observed)}

\item{metric}{string with metric of interest, e.g. 'mean_frac'}
}
\value{
dataframe of single row with 'source_target', 'mean_obs', 'overall_mean_shuffled' and 'pval'
}
\description{
Compute p-val permutation testing
}
