% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_spatial.R
\name{compute_shuffled_autocorr}
\alias{compute_shuffled_autocorr}
\title{spatialCrossCorTest}
\usage{
compute_shuffled_autocorr(i, x, y, w)
}
\arguments{
\item{i}{iteration (numeric)}

\item{x}{vector with expression of gene}

\item{y}{vector with expression of gene}

\item{w}{adjacency matrix}
}
\value{
dataframe with iteration and SCC
}
\description{
Adaptation of spatialCrossCorTest from MERINGUE.
}
