% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_compute_total_facetgrid_dimensions.R
\name{compute_total_facetgrid_dimensions}
\alias{compute_total_facetgrid_dimensions}
\title{Compute total facetgrid dimensions}
\usage{
compute_total_facetgrid_dimensions(
  n_plots,
  width = 5,
  height = 5,
  n_plots_per_row = 5
)
}
\arguments{
\item{n_plots}{totla number of facets (plots)}

\item{width}{width of single facet plot (inches)}

\item{height}{width of single facet plot (inches)}

\item{n_plots_per_row}{number of plots per row, i.e. number of columns}
}
\value{
list with \code{width}, \code{height} & \code{nrows}
}
\description{
Compute total facetgrid dimensions
}
