% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_h5ad_to_seurat.R
\name{convert_h5ad_to_seurat}
\alias{convert_h5ad_to_seurat}
\title{Convert H5ad Anndata object to Seurat object}
\usage{
convert_h5ad_to_seurat(h5ad_path, metadata_df = NULL)
}
\arguments{
\item{h5ad_path}{path to h5ad data}

\item{metadata_df}{dataframe where rownames are the barcodes/cell_ids that are present in the anndata object}
}
\value{
Seurat object
}
\description{
Converts a h5ad file containing an AnnData object (Python) to Seurat object. Due to compatibility issues, only the assay data is converted, metadata needs to be added manually using the \code{metadata_df}. Metadata from AnnData can be exported in Python using the snippet below
"""
import scanpy as sc
h5ad = "path-to-h5ad-file"
anndata = sc.read_h5ad(h5ad)
anndata.obs.to_csv("path-for-csv")
"""
}
