% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_list_to_df.R
\name{convert_list_to_df}
\alias{convert_list_to_df}
\title{Convert list to dataframe}
\usage{
convert_list_to_df(ls_vec)
}
\arguments{
\item{ls_vec}{list of vectors}
}
\value{
dataframe
}
\description{
Convert list of vectors with unequal lengths to dataframe
}
