% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir}
\alias{create_dir}
\title{Create directory}
\usage{
create_dir(dir_path, verbose = TRUE)
}
\arguments{
\item{dir_path}{Path to directory to be created}

\item{verbose}{log messages}
}
\description{
Create directory if it does not exist
}
\examples{
\dontrun{
create_dir("/Users/johndoe/my_new_dir")
}
}
