% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{create_hm}
\alias{create_hm}
\title{Create Heatmap using ComplexHeatmap}
\usage{
create_hm(matrix, cell_width = 4, cell_height = 4, ...)
}
\arguments{
\item{matrix}{matrix}

\item{cell_width}{with of cell in 'mm'}

\item{cell_height}{height of cell in 'mm'}
}
\value{
hm
}
\description{
Create Heatmap using ComplexHeatmap
}
