% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial.R
\name{create_lookup_table}
\alias{create_lookup_table}
\title{Create annotation lookup table}
\usage{
create_lookup_table(label, annot, label_name)
}
\arguments{
\item{label}{label}

\item{annot}{dataframe containing the required columns: 'Cluster' and 'Label'}

\item{label_name}{name for column with labels, e.g. Label}
}
\description{
Creates a look up table for annotating seurat object (Xenium) based on manually annotated clusters. Especially useful when multiple clusters have the same label
}
