% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_spatial.R
\name{create_polygon}
\alias{create_polygon}
\title{create polygons}
\usage{
create_polygon(x, cell_boundaries)
}
\arguments{
\item{x}{cell_id}

\item{cell_boundaries}{dataframe with at least the required columns: 'vertex_x', 'vertex_y' and 'cell_id'}
}
\value{
dataframe with 'vertex_x', 'vertex_y' and 'cell_id'
}
\description{
create polygons
}
