% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_scripts.R
\name{create_timestamped_logfile}
\alias{create_timestamped_logfile}
\title{Create timestamped logfile path}
\usage{
create_timestamped_logfile(log_file = NULL, log_dir = "logs")
}
\arguments{
\item{log_file}{(default = 'log')}

\item{log_dir}{(default = 'logs')}
}
\value{
string with path to logfile that needs to be created
}
\description{
Create a logfile path of format <log_dir>/\if{html}{\out{<timestamp>}}__<log_file>.log
}
