% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_values.R
\name{cutoff_values}
\alias{cutoff_values}
\title{Cut-off values by percentile}
\usage{
cutoff_values(vec, prob = 0.95)
}
\arguments{
\item{vec}{vector with numerical values}

\item{prob}{probability to use for cut-off}
}
\value{
vector
}
\description{
Cut-off values by percentile
}
