% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{default_theme}
\alias{default_theme}
\title{Standard theme for plots}
\usage{
default_theme(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
ggplot2 theme
}
\description{
Standard theme for plots
}
