% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{elbow_method_wrapper}
\alias{elbow_method_wrapper}
\title{Perform Elbow method}
\usage{
elbow_method_wrapper(seurat_object, ndims = 50, reduction = "pca")
}
\arguments{
\item{seurat_object}{seurat object}

\item{ndims}{Number of PCs to plot (default = 50)}

\item{reduction}{reduction to use for ElbowPlot (default = 'pca')}
}
\value{
list with plot and number of optimal npcs
}
\description{
Perform Elbow method
}
