% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format_p_value.R
\name{explode_p}
\alias{explode_p}
\title{Explode p-value}
\usage{
explode_p(p)
}
\arguments{
\item{p}{p-value}
}
\value{
named list with each item a single value
}
\description{
Dissects p-value to be used in \code{format_p_value}
}
