% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustree_extract_default_legend.R
\name{extract_default_clustree_legend}
\alias{extract_default_clustree_legend}
\title{Extract clustree legend}
\usage{
extract_default_clustree_legend(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
ggplot2 legend
}
\description{
Extracts the 'Cluster size', 'Sample count' and 'In-proportion' legend from a clustree.
}
