% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coloc_visium.R
\name{find_neighbouring_spots}
\alias{find_neighbouring_spots}
\title{Find neighbouring spots}
\usage{
find_neighbouring_spots(adj_table, spot_barcode)
}
\arguments{
\item{adj_table}{adjacency table}

\item{spot_barcode}{barcode of spot to select}
}
\value{
vector with barcodes of neighbors
}
\description{
Find neighbouring spots
}
\references{
Adapted from Greenwald 2024 (https://github.com/tiroshlab/Spatial_Glioma)
}
