% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coloc_visium.R
\name{find_neighbours_in_window}
\alias{find_neighbours_in_window}
\title{Find neighbours in window}
\usage{
find_neighbours_in_window(adj_table, radius)
}
\arguments{
\item{adj_table}{adjacency table}

\item{radius}{radius to use}
}
\value{
matrix
}
\description{
Find neighbours in window
}
\references{
Adapted from Greenwald 2024 (https://github.com/tiroshlab/Spatial_Glioma)
}
