% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_coloc_visium.R
\name{find_neighbours_in_window_around_spot}
\alias{find_neighbours_in_window_around_spot}
\title{Find neighbours in window around spot}
\usage{
find_neighbours_in_window_around_spot(adj_table, radius, spot_barcode)
}
\arguments{
\item{adj_table}{adjacency table}

\item{radius}{radius to use}

\item{spot_barcode}{barcode of spot to look at}
}
\value{
vector with barcodes of neighbors
}
\description{
Find neighbours in window around spot
}
\references{
Adapted from Greenwald 2024 (https://github.com/tiroshlab/Spatial_Glioma)
}
