% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format_number_as_string.R
\name{format_number_as_string}
\alias{format_number_as_string}
\title{Format number as string}
\usage{
format_number_as_string(x, n_dec = 2)
}
\arguments{
\item{x}{number/float}

\item{n_dec}{number of decimals to use}
}
\description{
Format number as string
}
