% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_format_p_value.R
\name{format_p_value}
\alias{format_p_value}
\title{Format p-value}
\usage{
format_p_value(p, smallest_p = 2.2e-16)
}
\arguments{
\item{p}{p-value}

\item{smallest_p}{smallest p value to display}
}
\value{
float
}
\description{
Format p-value for plots, reformat to 'A x 10^exp'
}
