% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_scientific}
\alias{format_scientific}
\title{Format large numbers with scientific notation}
\usage{
format_scientific(value)
}
\arguments{
\item{value}{the string or number to be formatted}
}
\description{
Format large numbers with scientific notation
}
\references{
https://stackoverflow.com/a/24241954
}
