% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaitiLabUtils-package.R
\docType{package}
\name{GaitiLabUtils-package}
\alias{GaitiLabUtils}
\alias{GaitiLabUtils-package}
\title{GaitiLabUtils: General (utility) functions for R-projects.}
\description{
Package containing functions that can be used in various contexts, developed for in-house use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gaitilab.github.io/GaitiLabUtils/}
}

}
\author{
\strong{Maintainer}: Joan Kant \email{Joan.Kant@uhn.ca} (\href{https://orcid.org/0009-0007-1837-5854}{ORCID})

Authors:
\itemize{
  \item Benson Wu
  \item Yiyan Wu
}

}
\keyword{internal}
