% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_scripts.R
\name{generate_logfile_name}
\alias{generate_logfile_name}
\title{Generate name for log}
\usage{
generate_logfile_name(logfile_name = NULL)
}
\arguments{
\item{logfile_name}{filename to save log, without file extension (default=NULL)}
}
\value{
string with filename
}
\description{
Generate name for logfile either returning default \if{html}{\out{<timestamp>}}\strong{log.log or \if{html}{\out{<timestamp>}}}\if{html}{\out{<Rscript>}}.log if run from command line
}
