% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_get_bracket_height.R
\name{get_bracket_height}
\alias{get_bracket_height}
\title{Set bracket height for p-values}
\usage{
get_bracket_height(df, col, metric, fct = 1.05)
}
\arguments{
\item{df}{dataframe}

\item{col}{name of column with groups (x-axis)}

\item{metric}{variable that is measured (y-axis)}

\item{fct}{factor to determine height}
}
\value{
dataframe with bracket heights for each level in \code{col}
}
\description{
Set bracket height for p-values
}
