% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_get_bracket_height.R
\name{get_bracket_height_rolling}
\alias{get_bracket_height_rolling}
\title{Dynamically set bracket heights}
\usage{
get_bracket_height_rolling(height, n, offset = 3, vert_spacing_fct = 6)
}
\arguments{
\item{height}{base height, e.g. max value in plot}

\item{n}{number of brackets, i.e. number of statistical tests performed}

\item{offset}{space between base height and first bracket}

\item{vert_spacing_fct}{factor to use for spacing between brackets}
}
\value{
numerical vector of length \code{n}
}
\description{
Compute bracket heights for plots with statistical tests, with brackets to show p-values. Here, the heights of the brackets is computed in a dynamic way, where \code{height} should be the value of the largest value in the plot, then \code{offset} is added to this height and offset controls the space between brackets, to prevent overlap.
}
