% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get color palette from Excel file}
\usage{
get_color_palette(name, excel_filepath, sheet_name = "color_palettes")
}
\arguments{
\item{name}{name of color palette}

\item{excel_filepath}{path to Excel file containing the color palette(s). A color palette is defined by two columns: (1) \if{html}{\out{<name>}}, (2) \if{html}{\out{<name>}}_color.}

\item{sheet_name}{the name of the sheet in the Excel file containing the color palette(s) (default='color_palettes')}
}
\value{
named list with the colors (\if{html}{\out{<name>}}_color) and names as \if{html}{\out{<name>}}
}
\description{
Get color palette from Excel file
}
