% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_spatial.R
\name{get_density}
\alias{get_density}
\title{Get density of points in 2 dimensions.}
\usage{
get_density(x, y, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{n}{Create a square n by n grid to compute density.}
}
\value{
The density within each square.
}
\description{
Get density of points in 2 dimensions.
}
