% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcs.R
\name{get_pcs}
\alias{get_pcs}
\title{Function for getting get_pcs}
\usage{
get_pcs(
  seurat_obj,
  reduc = "pca",
  method = c("elbow", "jackstraw")[1],
  cumu_filter = 90,
  perc_filter = 5,
  change_of_variation_perc = 0.05,
  n_dim = 30,
  dim_filter = 5^(-15)
)
}
\arguments{
\item{seurat_obj}{seurat object}

\item{reduc}{layer to use for 'reduction' (default='pca')}

\item{method}{method to use 'elbow' or 'jackstraw'
Parameters associated with 'elbow' method}

\item{cumu_filter}{PC exhibits cumulative percent greater than \code{cumu_filter}\% (default=90)}

\item{perc_filter}{PC exhibits associated with \%-variation < \code{perc_filter} (default=5)}

\item{change_of_variation_perc}{Last point where change of \% of variation is more than X\% used in 'elbow method' (default=0.05;0.05\%)
Parameters associated with 'jackstraw' method}

\item{n_dim}{number of dimensions to use for 'jackstraw' method (default=30)}

\item{dim_filter}{filter for JackStraw score and getting the optimal number of PCs used in the 'jackstraw' method (default=5^(-15))}
}
\description{
Getting the number of optimal PCs to use, either determined using the 'elbow' (faster) or 'jackstraw' method (slower).
Method 1: 'elbow'
\enumerate{
\item The point where the PCs only contribute 5\% of standard deviation and the principal components cumulatively contribute 90\% of the standard deviation.
\item The point where the percent change in variation between the consecutive PCs < 0.05\%.
}
}
\references{
https://github.com/tiroshlab/Spatial_Glioma/blob/main/Module1_MP.R (jackstraw implementation)

https://hbctraining.github.io/scRNA-seq/lessons/elbow_plot_metric.html (elbow implementation)
}
