% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser_setup_default.R,
%   R/log4r_init_logging.R
\name{init_obj_logging}
\alias{init_obj_logging}
\title{Set up object logging}
\usage{
init_obj_logging(log_file = NULL)

init_obj_logging(log_file = NULL)
}
\arguments{
\item{log_file}{path to log file (optional)}
}
\value{
logger object

logger object
}
\description{
Set up object logging

Set up object logging
}
\examples{
\dontrun{
logobj <- init_obj_logger()
}
\dontrun{
logobj <- init_obj_logger()
}
}
