% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_spatial.R
\name{inspect_neighbors}
\alias{inspect_neighbors}
\title{Inspect Neighbors}
\usage{
inspect_neighbors(cell_id, nn_id_list, group_varname, df_with_labels)
}
\arguments{
\item{cell_id}{cell_id for which you want to assess its neighbors}

\item{nn_id_list}{output from dbscan::frNN()}

\item{group_varname}{variable containing your labels for the cell_ids, e.g. cell type}

\item{df_with_labels}{dataframe with 'cell_ids' and your 'group_varname'}
}
\value{
return dataframe with the columns: labels_of_neighbors, cell_id, n_total, and n_cells
}
\description{
Determine the number of neighbors per cell type for a cell of interest
}
