% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_path}
\alias{is_valid_path}
\title{Check whether a path is valid}
\usage{
is_valid_path(
  path,
  expected_type = c("file", "dir")[1],
  required_file_extension = NULL
)
}
\arguments{
\item{path}{character string to be checked}

\item{expected_type}{'file' or 'dir'}

\item{required_file_extension}{in case expected_type == "file", you can also check for a specific file extension (case insensitive)}
}
\value{
logical
}
\description{
Check whether a path is valid
}
