% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Compute Jaccard similarity}
\usage{
jaccard_similarity(vec1, vec2)
}
\arguments{
\item{vec1}{vector}

\item{vec2}{vector}
}
\value{
jaccard similarity
}
\description{
Compute Jaccard similarity
}
