% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{LMM_test}
\alias{LMM_test}
\title{perform a nested t-test using LMMs.}
\usage{
LMM_test(data_df, response, condition, latent_vars, verbose = TRUE)
}
\arguments{
\item{data_df}{data frame containing data}

\item{response}{name of column in data_df that corresponds to the response variable}

\item{condition}{name of column in data_df that corresponds to condition of interest}

\item{latent_vars}{vector containing columns in data_df that correspond to variables that describe the nested structure of the data e.g. Patient}

\item{vebose}{(default=TRUE)}
}
\value{
p-value derived from likelihood ratio test of model incorporating condition vs null model
}
\description{
perform a nested t-test using LMMs.
}
