% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log4r_init_logging.R
\name{log_debug}
\alias{log_debug}
\title{Logging functions: log_debug}
\usage{
log_debug(...)
}
\arguments{
\item{...}{message}
}
\description{
Logging functions: log_debug
}
\examples{
\dontrun{
log_debug("Hello world!")
}
}
