% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser_setup_default.R,
%   R/log4r_init_logging.R
\name{log_object}
\alias{log_object}
\title{Log an object}
\usage{
log_object(...)

log_object(...)
}
\arguments{
\item{...}{message}
}
\description{
Log an object

Log an object
}
