% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dfs.R
\name{merge_dfs}
\alias{merge_dfs}
\title{Merge list of dataframes}
\usage{
merge_dfs(list_of_dfs, by, all.x = TRUE)
}
\arguments{
\item{list_of_dfs}{list of dataframes for merging}

\item{by}{merge on key (Default = i)}

\item{all.x}{merge on x (left dataframe)}
}
\description{
Merge list of dataframes
}
