% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scaling.R
\name{min_max_scaling}
\alias{min_max_scaling}
\title{Min-max scaling}
\usage{
min_max_scaling(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
single numerical value
}
\description{
Min-max scaling
}
