% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{normalizedata_wrapper}
\alias{normalizedata_wrapper}
\title{Wrapper for normalization}
\usage{
normalizedata_wrapper(
  seurat_object,
  assay = "RNA",
  apply_sct_transform = FALSE
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{assay}{RNA assay, either 'RNA.008um' (when created with Load10X_Spatial) or 'RNA' (default = "RNA")}

\item{apply_sct_transform}{logical indicating whether to apply sctransform (default = 0)}
}
\value{
seurat object
}
\description{
takes a seurat object and performs 1) Log-normalization and 2) SCTransform
}
