% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_one_sample_perm_test_pval.R
\name{one_sample_perm_test_pval}
\alias{one_sample_perm_test_pval}
\title{Compute p-val for one-sample permutation test}
\usage{
one_sample_perm_test_pval(
  null_dist,
  true_val,
  alternative = c("less", "greater", "two.sided")[1]
)
}
\arguments{
\item{null_dist}{vector with the null distribution obtained by shuffling/randomizing (without replacement)}

\item{true_val}{observed value}

\item{alternative}{alternative hypothesis, 'less', 'greater' or 'two.sided' (default = 'less')}
}
\description{
Compute p-value from null distribution and observed measurement (single value)
}
\references{
adapted from http://users.stat.umn.edu/~helwig/notes/perm-Notes.pdf (slide 18)
}
