% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_order_stacked_barplot.R
\name{order_stacked_barplot}
\alias{order_stacked_barplot}
\title{Order stacked barplot}
\usage{
order_stacked_barplot(df_wide, order_by, id, pc = 1)
}
\arguments{
\item{df_wide}{dataframe in wide format, containing at least the columns \code{order_by} and \code{id}, in addition with columns containing the measurements/values for different levels (categorical) of a variable.}

\item{order_by}{vector containing the names in \code{df_wide} to order by}

\item{id}{name of column in \code{df_wide} containing the ids, these are the bars in barplot}

\item{pc}{principal component to use for ordering}
}
\value{
vector with order for \code{id}
}
\description{
Order stacked barplot
}
