% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_coefficient.R
\name{overlap_coefficient}
\alias{overlap_coefficient}
\title{Compute overlap coefficient}
\usage{
overlap_coefficient(vec1, vec2)
}
\arguments{
\item{vec1}{character vector}

\item{vec2}{character vector}
}
\value{
single numerical value
}
\description{
Compute overlap coefficient
}
