% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser_setup_default.R
\name{params_ls_to_df}
\alias{params_ls_to_df}
\title{params_ls_to_df}
\usage{
params_ls_to_df(params_list)
}
\arguments{
\item{params_list}{list of parameters}
}
\value{
dataframe of parameters (2 columns)
}
\description{
Convert list of parameters to dataframe
}
