% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_pretty_breaks_unexpanded.R
\name{pretty_breaks_unexpanded}
\alias{pretty_breaks_unexpanded}
\title{Pretty unexpanded limits}
\usage{
pretty_breaks_unexpanded(x, n = 5)
}
\arguments{
\item{x}{vector of numeric values}

\item{n}{number of breaks (default=5)}
}
\description{
Set 'pretty' breaks based on vector of values. Can be used for \verb{breaks=} argument in ggplot2() functions. Alternative for scales::pretty_breaks() #nolint
}
\references{
https://stackoverflow.com/a/67865447/19429268
}
