% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_pretty_limits.R
\name{pretty_limits}
\alias{pretty_limits}
\title{Pretty limits}
\usage{
pretty_limits(x, n = 5)
}
\arguments{
\item{x}{vector of numeric values}

\item{n}{number of breaks (default=5)}
}
\description{
Set 'pretty' limits based on vector of values. Can be used for \verb{limits=} argument in ggplot2() functions #nolint
}
\references{
https://stackoverflow.com/a/67865447/19429268
}
