% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log4r_set_logfile.R
\name{set_logfile}
\alias{set_logfile}
\title{Set logfile}
\usage{
set_logfile(args, use_logs_subdir = TRUE)
}
\arguments{
\item{args}{list of arguments from argparser, should contain at least \code{log_dir}, \code{output_dir} and \code{log_file}}

\item{use_logs_subdir}{use a "logs" subdirectory}
}
\description{
Set logfile
}
