% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser_setup_default.R
\name{setup_default_argparser}
\alias{setup_default_argparser}
\title{initialize argparser}
\usage{
setup_default_argparser(
  description = "",
  default_output_dir = "output",
  default_log_file = NULL,
  default_log_dir = NULL
)
}
\arguments{
\item{description}{Description for script}

\item{default_log_file}{name of logfile without extension (default=NULL)}

\item{default_log_dir}{directory where logfile needs to be saved (default="output")}

\item{default_output}{Default output directory (default="output")}
}
\value{
parser object
}
\description{
Setup default argparser
}
\details{
Set up a default argument parser, with two default arguments: log_level and output_dir
}
\examples{
\dontrun{
setup_default_argparser(description = "Example")
}
}
