% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorensen_dice_index.R
\name{sorensen_dice_index}
\alias{sorensen_dice_index}
\title{Compute Sorenson-Dice index}
\usage{
sorensen_dice_index(vec1, vec2)
}
\arguments{
\item{vec1}{character vector}

\item{vec2}{character vector}
}
\value{
single numerical value
}
\description{
Compute Sorenson-Dice index
}
