% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_spatial.R
\name{visualize_transcripts}
\alias{visualize_transcripts}
\title{Visualize transcripts}
\usage{
visualize_transcripts(
  gene = NULL,
  transcripts,
  n_bins = 100,
  output_dir = ".",
  save_plot = TRUE,
  transcript_size = 0.1
)
}
\arguments{
\item{gene}{gene name}

\item{transcripts}{dataframe with the following required columns x_location, y_location and feature_name}

\item{n_bins}{number of bins to use for KDE2D (default = 100)}

\item{output_dir}{output directory for saving (default = ".")}

\item{save_plot}{whether to save the plot (default = TRUE)}

\item{transcript_size}{Size of transcript (dot) on plot (default = 0.1)}
}
\description{
visualize transcripts of a gene and color by density estimates
}
