% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_versions_yml.R
\name{write_versions_yml}
\alias{write_versions_yml}
\title{Write versions yml}
\usage{
write_versions_yml(list_of_pkgs = NULL, task_id = NULL, outdir = NULL)
}
\arguments{
\item{list_of_pkgs}{list (vector) of packages of interest (default=NULL). If Null uses pacman::p_loaded()  as list_of_pkgs}

\item{task_id}{(default=NULL)}

\item{outdir}{directory to save the 'versions.yml' (default=NULL)}
}
\description{
Creates a yml file with the R version and versions of the packages of interest, e.g. loaded packages in a script. Useful for Nextflow pipelines
}
