% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{create_hm}
\alias{create_hm}
\title{Plot heatmap + save}
\usage{
create_hm(matrix, cell_width = 4, cell_height = 4, is_full = TRUE, ...)
}
\arguments{
\item{matrix}{matrix}

\item{cell_width}{height of cell in mm, default=0.1 (square)}

\item{cell_height}{width of cell in mm, default=0.1 (square)}

\item{is_full}{plot full matrix (default = TRUE)}
}
\value{
hm
}
\description{
Plot heatmap + save
}
