% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_elbow_method_wrapper.R
\name{elbow_method_wrapper}
\alias{elbow_method_wrapper}
\title{Perform Elbow method}
\usage{
elbow_method_wrapper(
  seurat_object,
  assay = "RNA",
  ndims = 50,
  reduction = "pca",
  var = 0.05
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{assay}{assay to use in Seurat object (default = 'RNA')}

\item{ndims}{Number of PCs to plot (default = 50)}

\item{reduction}{reduction to use for ElbowPlot (default = 'pca')}

\item{change_of_variation_perc}{Last point where change of \% of variation is more than X\%}
}
\value{
list with plot and number of optimal npcs
}
\description{
Perform Elbow method
}
