% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_scripts.R
\name{init_obj_logging}
\alias{init_obj_logging}
\title{Set up object logging}
\usage{
init_obj_logging(log_file = NULL)
}
\arguments{
\item{log_file}{path to log file (optional)}
}
\value{
logger object
}
\description{
Set up object logging
}
\examples{
\dontrun{
logobj <- init_obj_logger()
}
}
