% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_scripts.R
\name{log_error}
\alias{log_error}
\title{Logging functions: log_error}
\usage{
log_error(logr, ...)
}
\arguments{
\item{logr}{logger object}

\item{...}{message}
}
\description{
Logging functions: log_error
}
\examples{
\dontrun{
log_error("Hello world!")
}
}
