% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_scripts.R
\name{log_fatal}
\alias{log_fatal}
\title{Logging functions: log_fatal}
\usage{
log_fatal(logr, ...)
}
\arguments{
\item{logr}{logger object}

\item{...}{message}
}
\description{
Logging functions: log_fatal
}
\examples{
\dontrun{
log_fatal("Hello world!")
}
}
