% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argparser_setup_default.R,
%   R/log4r_init_logging.R
\name{log_fatal}
\alias{log_fatal}
\title{Logging functions: log_fatal}
\usage{
log_fatal(...)

log_fatal(...)
}
\arguments{
\item{...}{message}

\item{logr}{logger object}
}
\description{
Logging functions: log_fatal

Logging functions: log_fatal
}
\examples{
\dontrun{
log_fatal("Hello world!")
}
\dontrun{
log_fatal("Hello world!")
}
}
