% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_scripts.R
\name{log_info}
\alias{log_info}
\title{Logging functions: log_info}
\usage{
log_info(logr, ...)
}
\arguments{
\item{logr}{logger object}

\item{...}{message}
}
\description{
Logging functions: log_info
}
\examples{
\dontrun{
log_info("Hello world!")
}
}
