% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log4r_init_logging.R
\name{log_info}
\alias{log_info}
\title{Logging functions: log_info}
\usage{
log_info(...)
}
\arguments{
\item{...}{message}
}
\description{
Logging functions: log_info
}
\examples{
\dontrun{
log_info("Hello world!")
}
}
