% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log4r_init_logging.R
\name{log_warn}
\alias{log_warn}
\title{Logging functions: log.warn}
\usage{
log_warn(...)
}
\arguments{
\item{...}{message}
}
\description{
Logging functions: log.warn
}
\examples{
\dontrun{
log_warn("Hello world!")
}
}
